有向グラフの実装

+++ test/weighted_node_test.cpp
@@ -80,10 +80,10 @@ TEST(WeightedNodeTest, RemoveAdjacent) {
     WeightedNode node1;
     node1.setAdjacent(4, 1.5);
     node1.setAdjacent(5, 2.0);
-    EXPECT_EQ(node1.getAdjacents().size(), 2);
+    EXPECT_EQ(static_cast<int>(node1.getAdjacents().size()), 2);
     node1.removeAdjacent(5);
     const auto& adjacents = node1.getAdjacents();
-    EXPECT_EQ(adjacents.size(), 1);
+    EXPECT_EQ(adjacents.size(), static_cast<size_t>(1));
     EXPECT_FALSE(adjacents.contains(5));
 }
